refactor: update config component filenames to PascalCase
