+++ web/components/config/EditCustomStyles.tsx
@@ -15,9 +15,9 @@ import {
   STATUS_PROCESSING,
   STATUS_SUCCESS,
 } from '../../utils/input-statuses';
-import FormStatusIndicator from './form-status-indicator';
+import FormStatusIndicator from './FormStatusIndicator';
 
-import TextField, { TEXTFIELD_TYPE_TEXTAREA } from './form-textfield';
+import TextField, { TEXTFIELD_TYPE_TEXTAREA } from './TextField';
 import { UpdateArgs } from '../../types/config-section';
 
 const { Title } = Typography;
