+++ web/components/config/ResetYP.tsx
@@ -10,7 +10,7 @@ import {
   STATUS_PROCESSING,
   STATUS_SUCCESS,
 } from '../../utils/input-statuses';
-import FormStatusIndicator from './form-status-indicator';
+import FormStatusIndicator from './FormStatusIndicator';
 
 export default function ResetYP() {
   const { setMessage } = useContext(AlertMessageContext);
