+++ web/pages/admin/actions.tsx
@@ -1,7 +1,7 @@
 import { DeleteOutlined } from '@ant-design/icons';
 import { Button, Checkbox, Input, Modal, Space, Table, Typography } from 'antd';
 import React, { useContext, useEffect, useState } from 'react';
-import FormStatusIndicator from '../../components/config/form-status-indicator';
+import FormStatusIndicator from '../../components/config/FormStatusIndicator';
 import {
   API_EXTERNAL_ACTIONS,
   postConfigUpdateToAPI,
